created backend_ui.py for gradio functions
