+++ backend/alembic/versions/10ed6e0bd75f_renamed_openai_key_to_user.py
@@ -19,18 +19,8 @@ depends_on: Union[str, Sequence[str], None] = None
 
 
 def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    # op.add_column('user_data', sa.Column('name', sa.String(), nullable=False))
-    # op.create_unique_constraint(None, 'user_data', ['name'])
-    # op.drop_column('user_data', 'openai_key')
     op.alter_column("user_data", "openai_key", new_column_name="name")
-    # ### end Alembic commands ###
 
 
 def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    # op.add_column('user_data', sa.Column('openai_key', sa.VARCHAR(), nullable=False))
-    # op.drop_constraint(None, 'user_data', type_='unique')
-    # op.drop_column('user_data', 'name')
     op.alter_column("user_data", "name", new_column_name="openai_key")
-    # ### end Alembic commands ###
